Make header server-side
