+++ src/components/Header/Header.tsx
@@ -1,31 +1,16 @@
-'use client';
-
-import React, { useState } from 'react';
 import styles from './Header.module.scss';
 import EscapeHatch from './EscapeHatch/EscapeHatch';
-import Navigation from './Navigation/Navigation';
-import User from './User/User';
+import SearchSwitcher from './SearchSwitcher/SearchSwitcher';
 import ThemeSelector from './ThemeSelector/ThemeSelector';
-import Search from './SearchBar/Search';
+import User from './User/User';
 
-const Header = () => {
-  const [isSearchExpanded, setIsSearchExpanded] = useState(false);
-
-  const handleSearchIconClick = () => {
-    setIsSearchExpanded(!isSearchExpanded);
-  };
-
+const Header = async () => {
   return (
     <header className={styles.header}>
       <EscapeHatch />
-      {isSearchExpanded ? null : <Navigation />}
-      <div>
-        <Search
-          onSearchIconClick={handleSearchIconClick}
-          isExpanded={isSearchExpanded}
-        />
-      </div>
+      <SearchSwitcher />
       <ThemeSelector />
+      {/* @ts-expect-error Server Component */}
       <User />
     </header>
   );
